migrate

+++ database/migrations/2024_07_05_152999_create_reporte_fotografico_new_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('base', 100)->nullable();
             $table->string('foto', 100)->nullable();
-            $table->integer('codigo')->nullable();
+            $table->unsignedBigInteger('codigo')->nullable();
             $table->integer('estado')->nullable();
             $table->dateTime('fec_reg')->nullable();
             $table->integer('user_reg')->nullable();
